refactor: Move UserTextRequest to new file and make RequstType type
